rename SyncData to SyncStatus
